# openapi-spec.yaml
openapi: "3.0.3"
info:
  title: WireGuard RESTful API
  description: |
    Manage WireGuard VPN tunnels by RESTful manner.

    Supported features:

      * Manage device: create, update, and delete wireguard interface
      * Manage device's ip addresses: attache or detach ip addresses to the netowrk interface
      * Manage device's peers: create, update, and delete peers
      * Peer's QR code, for use in WireGuard & ForestVPN client

    ForestVPN client may be used as alternative client with enabled P2P technology over WireGuard tunnelling.
    Read more on https://forestvpn.com/
  termsOfService: https://forestvpn.com/terms/
  contact:
    name: ForestVPN
    url: https://forestvpn.com/
    email: support@forestvpn.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: "1.0"
...
servers:
  - url: '{scheme}://{host}/v1'
    variables:
      host:
        default: "example.com"
      scheme:
        enum:
          - 'https'
          - 'http'
        default: 'https'
security:
  - bearerAuth: [ ]
paths:
  /devices/:
    get:
      summary: Devices list
      operationId: ListDevices
      tags:
        - device
      parameters:
        - in: query
          name: per_page
          description: Number of WireGuard devices per page. Default is 100
          schema:
            type: number
        - in: query
          name: page
          description: Page number. Default is 0
          schema:
            type: number
      responses:
        '200':
          description: ok
          headers:
            Link:
              schema:
                type: string
                description: https://docs.github.com/en/rest/guides/traversing-with-pagination
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create new device
      operationId: CreateDevice
      tags:
        - device
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceCreateOrUpdateRequest"
      responses:
        201:
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        409:
          description: Device exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /devices/{name}/:
    get:
      summary: Get device info
      operationId: GetDevice
      tags:
        - device
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        404:
          description: not found
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete Device
      operationId: DeleteDevice
      tags:
        - device
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        204:
          description: No content
        404:
          description: Not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      summary: Update device
      operationId: UpdateDevice
      tags:
        - device
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceCreateOrUpdateRequest"
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        404:
          description: Not found
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /devices/{name}/options/:
    get:
      summary: Get device options
      operationId: GetDeviceOptions
      tags:
        - device
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceOptions"
        404:
          description: not found
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      summary: Update device's options
      operationId: UpdateDeviceOptions
      tags:
        - device
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceOptionsUpdateRequest"
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceOptions"
        404:
          description: Not found
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /devices/{name}/peers/:
    get:
      summary: Peers list
      operationId: ListDevicePeers
      tags:
        - device
      parameters:
        - in: path
          name: name
          description: Device's name
          required: true
          schema:
            type: string
        - in: query
          name: per_page
          description: Number of WireGuard device's peers per page. Default is 100
          schema:
            type: number
        - in: query
          name: page
          description: Page number. Default is 0
          schema:
            type: number
        - in: query
          name: q
          description: Search query
          schema:
            type: string
        - in: query
          name: sort
          description: Sort field
          schema:
            type: string
            enum:
              - pub_key
              - receive_bytes
              - -receive_bytes
              - transmit_bytes
              - -transmit_bytes
              - total_bytes
              - -total_bytes
              - last_handshake_time
              - -last_handshake_time
      responses:
        200:
          description: Ok
          headers:
            Link:
              schema:
                type: string
                description: https://docs.github.com/en/rest/guides/traversing-with-pagination
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Peer"
        404:
          description: Device not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create new device peer
      operationId: CreateDevicePeer
      tags:
        - device
      parameters:
        - in: path
          name: name
          description: Device's name
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeerCreateOrUpdateRequest"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Peer"
        404:
          description: Device not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /devices/{name}/peers/{urlSafePubKey}/:
    get:
      summary: Get device peer info
      operationId: GetDevicePeer
      tags:
        - device
      parameters:
        - in: path
          name: name
          description: Device's name
          required: true
          schema:
            type: string
        - in: path
          name: urlSafePubKey
          description: Peer's url safe public key
          required: true
          schema:
            type: string
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Peer"
        404:
          description: Peer or device not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      summary: Update device's peer
      operationId: UpdateDevicePeer
      tags:
        - device
      parameters:
        - in: path
          name: name
          description: Device's name
          required: true
          schema:
            type: string
        - in: path
          name: urlSafePubKey
          description: Peer's url safe public key
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeerCreateOrUpdateRequest"
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Peer"
        404:
          description: Peer or device not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete device's peer
      operationId: DeleteDevicePeer
      tags:
        - device
      parameters:
        - in: path
          name: name
          description: Device's name
          required: true
          schema:
            type: string
        - in: path
          name: urlSafePubKey
          description: Peer's url safe public key
          required: true
          schema:
            type: string
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Peer"
        404:
          description: Peer or device not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /devices/{name}/peers/{urlSafePubKey}/quick.conf:
    get:
      summary: Get device peer quick config
      operationId: GetDevicePeerQuickConfig
      tags:
        - device
      parameters:
        - in: path
          name: name
          description: Device's name
          required: true
          schema:
            type: string
        - in: path
          name: urlSafePubKey
          description: Peer's url safe public key
          required: true
          schema:
            type: string
      responses:
        200:
          description: Ok
          content:
            text/plain:
              schema:
                type: string
                format: binary
        404:
          description: Peer or device not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /devices/{name}/peers/{urlSafePubKey}/quick.conf.png:
    get:
      summary: Get device peer quick config QR code
      operationId: GetDevicePeerQuickConfigQRCodePNG
      tags:
        - device
      parameters:
        - in: path
          name: name
          description: Device's name
          required: true
          schema:
            type: string
        - in: path
          name: urlSafePubKey
          description: Peer's url safe public key
          required: true
          schema:
            type: string
        - in: query
          name: width
          description: QR code's width & height. Default is 256px.
          schema:
            type: string
      responses:
        200:
          description: Ok
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        404:
          description: Peer or device not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Authorization token
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
          example: device_does_not_exist
        message:
          type: string
          description: Error's short description
          example: Device "wg0" does'n exist
        detail:
          type: string
          description: Error's detail description
...